requirements.txt
celery==5.3.4
Flask==3.0.0
flask_wtf==1.2.1
fuzzywuzzy==0.18.0
llama_cpp_python==0.1.83
openpyxl==3.1.2
pandas==2.0.3
rapidfuzz==3.4.0
WTForms==3.1.1
